+++ src/tokens.js
@@ -1264,6 +1264,190 @@ module.exports = [
         ],
     },
     {
+        name: 'Duration',
+        tokens: [
+            {
+                id: '1',
+                type: 'string',
+                platforms: {
+                    scss: {
+                        name: 'tp-duration__1',
+                        value: '75ms',
+                    },
+                    javascript: {
+                        name: 'tpDuration1',
+                        value: '75ms',
+                    },
+                    ios: {
+                        name: 'duration1',
+                        value: '0.075',
+                    },
+                    android: {
+                        name: 'tp_duration_1',
+                        value: '75',
+                    },
+                },
+            },
+            {
+                id: '2',
+                type: 'string',
+                platforms: {
+                    scss: {
+                        name: 'tp-duration__2',
+                        value: '150ms',
+                    },
+                    javascript: {
+                        name: 'tpDuration2',
+                        value: '150ms',
+                    },
+                    ios: {
+                        name: 'duration2',
+                        value: '0.15',
+                    },
+                    android: {
+                        name: 'tp_duration_2',
+                        value: '150',
+                    },
+                },
+            },
+            {
+                id: '3',
+                type: 'string',
+                platforms: {
+                    scss: {
+                        name: 'tp-duration__3',
+                        value: '200ms',
+                    },
+                    javascript: {
+                        name: 'tpDuration3',
+                        value: '200ms',
+                    },
+                    ios: {
+                        name: 'duration3',
+                        value: '0.2',
+                    },
+                    android: {
+                        name: 'tp_duration_3',
+                        value: '200',
+                    },
+                },
+            },
+            {
+                id: '4',
+                type: 'string',
+                platforms: {
+                    scss: {
+                        name: 'tp-duration__4',
+                        value: '250ms',
+                    },
+                    javascript: {
+                        name: 'tpDuration4',
+                        value: '250ms',
+                    },
+                    ios: {
+                        name: 'duration4',
+                        value: '0.25',
+                    },
+                    android: {
+                        name: 'tp_duration_4',
+                        value: '250',
+                    },
+                },
+            },
+            {
+                id: '5',
+                type: 'string',
+                platforms: {
+                    scss: {
+                        name: 'tp-duration__5',
+                        value: '300ms',
+                    },
+                    javascript: {
+                        name: 'tpDuration5',
+                        value: '300ms',
+                    },
+                    ios: {
+                        name: 'duration5',
+                        value: '0.3',
+                    },
+                    android: {
+                        name: 'tp_duration_5',
+                        value: '300',
+                    },
+                },
+            },
+            {
+                id: '6',
+                type: 'string',
+                platforms: {
+                    scss: {
+                        name: 'tp-duration__6',
+                        value: '350ms',
+                    },
+                    javascript: {
+                        name: 'tpDuration6',
+                        value: '350ms',
+                    },
+                    ios: {
+                        name: 'duration6',
+                        value: '0.35',
+                    },
+                    android: {
+                        name: 'tp_duration_6',
+                        value: '350',
+                    },
+                },
+            },
+        ],
+    },
+    {
+        name: 'Easing',
+        tokens: [
+            {
+                id: 'ease-in',
+                type: 'string',
+                platforms: {
+                    scss: {
+                        name: 'tp-ease__in',
+                        value: 'cubic-bezier(0.50, 0, 1, 1)',
+                    },
+                    javascript: {
+                        name: 'tpEaseIn',
+                        value: 'cubic-bezier(0.50, 0, 1, 1)',
+                    },
+                },
+            },
+            {
+                id: 'ease-out',
+                type: 'string',
+                platforms: {
+                    scss: {
+                        name: 'tp-ease__out',
+                        value: 'cubic-bezier(0, 0, 0.40, 1)',
+                    },
+                    javascript: {
+                        name: 'tpEaseOut',
+                        value: 'cubic-bezier(0, 0, 0.40, 1)',
+                    },
+                },
+            },
+            {
+                id: 'ease-in-out',
+                type: 'string',
+                platforms: {
+                    scss: {
+                        name: 'tp-ease__in-out',
+                        value: 'cubic-bezier(0.45, 0, 0.40, 1)',
+                    },
+                    javascript: {
+                        name: 'tpEaseInOut',
+                        value: 'cubic-bezier(0.45, 0, 0.40, 1)',
+                    },
+                },
+            },
+        ],
+    },
+    {
         name: 'Font Family',
         tokens: [
             {